+++ CHANGELOG.md
@@ -7,7 +7,9 @@ Versioning](http://semver.org/)!
 
 ## [Unreleased]
 ### Changed
-- Require HLint >=2.0.8 or HLint < 2 and >= 1.9.25
+- Require HLint >=2.0.8 or HLint < 2 and >= 1.9.25 because the extension now
+  feeds document contents to HLint on standard input which only works in the
+  aforementioned versions.
 
 ## 0.3.1 – 2017-05-22
 ### Added